Move pylxd.operation.LXDOperation to pylxd.deprecated.operation
