+++ pylxd/deprecated/api.py
@@ -20,7 +20,7 @@ from pylxd.deprecated import container
 from pylxd.deprecated import hosts
 from pylxd.deprecated import image
 from pylxd.deprecated import network
-from pylxd import operation
+from pylxd.deprecated import operation
 from pylxd import profiles
